requirements.txt
Flask==0.10.1
google-api-python-client==1.2
GoogleAppEngineCloudStorageClient==1.9.0.0
httplib2==0.9
simplekv==0.8.4
Flask-KVSession==0.4
